Remove `perform_restore` option from configs.
The satellite-clone playbook was also used to quickly install a sat 6.1 or sat 6.2 box without actually restoring backup data. But this beats the purpose of this playbook and this playbook should stay simple and support just cloning. So perform_restore option will be removed.  A new playbook needs to be created in future if there is an intent to support plain install of Satellite without backup data.

Fixes #97