Use local vars files for user input
Fixes #104
Users will edit the copied sample yml files in the top level of our
project and edit them as necessary. The vars files in the roles
of satellite-hostname and satellite-clone are used to set defaults
which are overrided by any other vars files or vars passed into
the playbook.