Actually drop mongo database
The echo command we were using wasn't actually dropping the mongo
database, but rather echoing 'db.dropDatabase();' | mongo pulp_database

This resulted in data being restored on an existing database.

So the mongo database stayed during restore. We didn't notice many issues
because not much is in the mongo database on a clean satellite install.

Fixes #259