Disables hammer-cli-csv module and ignore warnings in hammer commands
This disables the hammer-cli-csv module by removing it's configuration file from the hammer config. This is an issue on some 6.6 installations, where the file is still present.

A fix could still be needed in the installer, but this commit will fix things for any current backups by ensuring the file is removed.

This also updates the disassociate_capsules script to exclude warnings for all hammer commands where we are parsing the output. It also DRYs up our parsing into one function to make updating this area easier.

You can test just the "disassociate_capsules" script on a Satellite that includes capsules associated with a lifecycle environment. The script will disassociate them and give instructions for reversal.