Remove support before 6.4 and don't require candlepin password
The cpdb update step is handled by the installer and is cached in a file now. The step was a workaround so we can remove it, which will fix the blank cp pw in answers file issue.

This commit also removes 6.4 support, updates docs, and removes misc no-longer-relevant items like cacheable facts.