Disable Smart Proxy Realm plugin
Similar to 5d94c6ee25eccff05c8f894217aed6245e186d34 this disables realm
integration on cloning. This is because the new host is usually not
enrolled in IPA so the files it tries to read are not present. That
causes clones to fail.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1965844