Capsule 6.10 support
The original "partial Satellite 6.10 support" change was missing support
for disassociating external capsules.
This change adds support for those, by checking for the
"Container_Gateway" feature, as the "Pulp Node" feature is gone with
Pulp3 and there is (currently) no better way to identify an *external*
capsule -- all of them have the Pulpcore feature.